Rename cqc provider cleaned column names file
