+++ tests/test_file_schemas.py
@@ -53,7 +53,7 @@ from utils.column_names.raw_data_files.ascwds_workplace_columns import (
 from utils.column_names.cleaned_data_files.cqc_pir_cleaned import (
     CqcPIRCleanedColumns as CQCPIRClean,
 )
-from utils.column_names.cleaned_data_files.cqc_provider_cleaned_values import (
+from utils.column_names.cleaned_data_files.cqc_provider_cleaned import (
     CqcProviderCleanedColumns as CQCPClean,
 )
 from utils.column_names.cleaned_data_files.cqc_location_cleaned import (
