+++ utils/column_names/cleaned_data_files/cqc_location_cleaned.py
@@ -3,7 +3,7 @@ from dataclasses import dataclass
 from utils.column_names.raw_data_files.cqc_location_api_columns import (
     NewCqcLocationApiColumns,
 )
-from utils.column_names.cleaned_data_files.cqc_provider_cleaned_values import (
+from utils.column_names.cleaned_data_files.cqc_provider_cleaned import (
     CqcProviderCleanedColumns as CQCPClean,
     CqcProviderCleanedValues as CQCPValues,
 )
