+++ utils/validation/validation_rules/providers_api_cleaned_validation_rules.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from utils.column_names.cleaned_data_files.cqc_provider_cleaned_values import (
+from utils.column_names.cleaned_data_files.cqc_provider_cleaned import (
     CqcProviderCleanedColumns as CQCPClean,
 )
 from utils.column_values.categorical_columns_by_dataset import (
